Move functions to top-level for memoization
